Add support for puppet 2.7.x
In the `pg_role` type different properties were derived from the
`Puppet::Property::Boolean` property class which was introduced in
puppet 3. As a consquence this code fails on puppet 2.7.x

Do not declare the properties as boolean anymore and use the standard
`newvalues` method to only allow the two states :true and :false.

The type can now be used on puppet 2.7.x as well as 3.x