Add requirements to the README file
The providers assume that they can run `psql` as the `postgres` user
to create/modify/destroy databases/roles without user interaction and
without the requirement to enter a password. This means that the
authentication method for the `postgres` user has be either `trust`
(huge security risk) or `ident`.

This is now called out in the README file.